+++ opt/src/commonMain/kotlin/ir/platform/x64/CompilationUnit.kt
@@ -109,7 +109,7 @@ class CompilationUnit: CompiledModule, ObjModule(NameAssistant()) {
         is I32Value -> builder.long(data.i32)
         is U32Value -> builder.long(data.u32)
         is I16Value -> builder.short(data.i16)
-        is U16Value -> builder.short(data.u16)
+        is U16Value -> builder.short(data.u16.toShort())
         is I8Value  -> builder.byte(data.i8)
         is U8Value  -> builder.byte(data.u8.toByte())
         is F32Value -> {
